# contracts/openapi/infraflow-api-v1.openapi.json
{"openapi":"3.1.0","info":{"title":"InfraFlow Operations API","description":"REST API for incidents, work orders and controlled operations workflows.","license":{"name":"MIT"},"version":"v1"},"servers":[{"url":"/"}],"security":[{"bearerAuth":[]}],"tags":[{"name":"Authentication","description":"JWT-based authentication endpoints"},{"name":"Agent sessions","description":"Provider-neutral advisory sessions; no LLM is called."},{"name":"Assets","description":"Read-only operational asset registry"},{"name":"Incidents","description":"Operations incident triage and workflow commands"},{"name":"Work orders","description":"Controlled work order drafting and inspection workflows"}],"paths":{"/api/v1/work-orders/{workOrderId}/start":{"post":{"tags":["Work orders"],"summary":"Start work on a ready work order","operationId":"startWorkOrder","parameters":[{"name":"workOrderId","in":"path","required":true,"schema":{"type":"string","pattern":"WO-\\d{4}-\\d{4}"}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/WorkOrderWorkflowResponse"}}}}}}},"/api/v1/work-orders/{workOrderId}/ready":{"post":{"tags":["Work orders"],"summary":"Move a drafted work order to ready","operationId":"moveWorkOrderToReady","parameters":[{"name":"workOrderId","in":"path","required":true,"schema":{"type":"string","pattern":"WO-\\d{4}-\\d{4}"}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/WorkOrderWorkflowResponse"}}}}}}},"/api/v1/work-orders/{workOrderId}/complete":{"post":{"tags":["Work orders"],"summary":"Complete a work order after verification","operationId":"completeWorkOrder","parameters":[{"name":"workOrderId","in":"path","required":true,"schema":{"type":"string","pattern":"WO-\\d{4}-\\d{4}"}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/WorkOrderWorkflowResponse"}}}}}}},"/api/v1/work-orders/drafts":{"post":{"tags":["Work orders"],"summary":"Draft a work order from an incident","description":"Creates a controlled work order draft from an existing incident context.","operationId":"draftWorkOrder","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/DraftWorkOrderRequest"}}},"required":true},"responses":{"201":{"description":"Work order draft created","content":{"application/json":{"schema":{"$ref":"#/components/schemas/WorkOrderResponse"}}}},"400":{"description":"Request validation failed","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"404":{"description":"Incident not found","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApiError"}}}}}}},"/api/v1/incidents":{"get":{"tags":["Incidents"],"summary":"Search incidents","description":"Returns a page of the incident queue filtered by free text and optional severity.","operationId":"searchIncidents","parameters":[{"name":"searchTerm","in":"query","description":"Free text filter for id, asset, location or title","required":false,"schema":{"type":"string","default":""}},{"name":"severity","in":"query","description":"Severity filter. Use All, Critical, High, Medium or Low.","required":false,"schema":{"type":"string","default":"All"}},{"name":"page","in":"query","description":"Zero-based page index.","required":false,"schema":{"type":"integer","format":"int32","default":0},"example":0},{"name":"size","in":"query","description":"Page size, clamped to 1-100.","required":false,"schema":{"type":"integer","format":"int32","default":20},"example":20},{"name":"sort","in":"query","description":"Sort as property,direction. Properties: reportedAt, title, severity, status, id.","required":false,"schema":{"type":"string","default":"reportedAt,desc"},"example":"reportedAt,desc"}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/IncidentPageResponse"}}}}}},"post":{"tags":["Incidents"],"summary":"Report a new incident","description":"Creates a new operations incident. The initial status is Open.","operationId":"createIncident","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateIncidentRequest"}}},"required":true},"responses":{"201":{"description":"Incident created","content":{"application/json":{"schema":{"$ref":"#/components/schemas/IncidentResponse"}}}},"400":{"description":"Request validation failed","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApiError"}}}}}}},"/api/v1/incidents/{incidentId}/start-response":{"post":{"tags":["Incidents"],"summary":"Start response workflow","description":"Moves an acknowledged incident into active response coordination.","operationId":"startIncidentResponse","parameters":[{"name":"incidentId","in":"path","required":true,"schema":{"type":"string","pattern":"INC-\\d{4}-\\d{4}"},"example":"INC-2026-0001"}],"responses":{"200":{"description":"Response workflow started","content":{"application/json":{"schema":{"$ref":"#/components/schemas/IncidentResponse"}}}},"404":{"description":"Incident not found","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"409":{"description":"Workflow rule rejected the command","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApiError"}}}}}}},"/api/v1/incidents/{incidentId}/resolve":{"post":{"tags":["Incidents"],"summary":"Resolve an incident","description":"Closes an incident after operational ownership and response have been verified. Admin only.","operationId":"resolveIncident","parameters":[{"name":"incidentId","in":"path","required":true,"schema":{"type":"string","pattern":"INC-\\d{4}-\\d{4}"},"example":"INC-2026-0001"}],"responses":{"200":{"description":"Incident resolved","content":{"application/json":{"schema":{"$ref":"#/components/schemas/IncidentResponse"}}}},"403":{"description":"Admin role is required","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"404":{"description":"Incident not found","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"409":{"description":"Workflow rule rejected the command","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApiError"}}}}}}},"/api/v1/incidents/{incidentId}/acknowledge":{"post":{"tags":["Incidents"],"summary":"Acknowledge an incident","description":"Marks an open incident as seen by an operator without resolving the physical issue.","operationId":"acknowledgeIncident","parameters":[{"name":"incidentId","in":"path","required":true,"schema":{"type":"string","pattern":"INC-\\d{4}-\\d{4}"},"example":"INC-2026-0001"}],"responses":{"200":{"description":"Incident acknowledged","content":{"application/json":{"schema":{"$ref":"#/components/schemas/IncidentResponse"}}}},"404":{"description":"Incident not found","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"409":{"description":"Workflow rule rejected the command","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApiError"}}}}}}},"/api/v1/auth/refresh":{"post":{"tags":["Authentication"],"summary":"Refresh an access token","operationId":"refresh","parameters":[{"name":"infraflow_refresh","in":"cookie","required":false,"schema":{"type":"string"}}],"responses":{"200":{"description":"New JWT token pair issued","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TokenResponse"}}}},"401":{"description":"Invalid refresh token","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApiError"}}}}},"security":[]}},"/api/v1/auth/logout":{"post":{"tags":["Authentication"],"summary":"Revoke the refresh token and clear its cookie","operationId":"logout","parameters":[{"name":"infraflow_refresh","in":"cookie","required":false,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK"}},"security":[]}},"/api/v1/auth/login":{"post":{"tags":["Authentication"],"summary":"Login with demo credentials","operationId":"login","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginRequest"}}},"required":true},"responses":{"200":{"description":"JWT token pair issued","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TokenResponse"}}}},"401":{"description":"Invalid credentials","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApiError"}}}}},"security":[]}},"/api/v1/work-orders":{"get":{"tags":["Work orders"],"summary":"List work orders","operationId":"listWorkOrders","responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/WorkOrderResponse"}}}}}}}},"/api/v1/work-orders/{workOrderId}":{"get":{"tags":["Work orders"],"summary":"Get work order by id","operationId":"getWorkOrder","parameters":[{"name":"workOrderId","in":"path","required":true,"schema":{"type":"string","pattern":"WO-\\d{4}-\\d{4}"},"example":"WO-2026-0001"}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/WorkOrderResponse"}}}}}}},"/api/v1/incidents/{incidentId}":{"get":{"tags":["Incidents"],"summary":"Get incident by id","operationId":"getIncident","parameters":[{"name":"incidentId","in":"path","required":true,"schema":{"type":"string","pattern":"INC-\\d{4}-\\d{4}"},"example":"INC-2026-0001"}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/IncidentResponse"}}}}}}},"/api/v1/assets":{"get":{"tags":["Assets"],"summary":"Search registered infrastructure assets","operationId":"searchAssets","parameters":[{"name":"searchTerm","in":"query","description":"Free text filter for id, name, type or location","required":false,"schema":{"type":"string","default":""}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/AssetResponse"}}}}}}}},"/api/v1/assets/{assetId}":{"get":{"tags":["Assets"],"summary":"Get asset by id","operationId":"getAsset","parameters":[{"name":"assetId","in":"path","required":true,"schema":{"type":"string","pattern":"[A-Z]{2,5}-[A-Z]{2,5}-\\d{3}"},"example":"TRF-NT-003"}],"responses":{"200":{"description":"Asset detail","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AssetResponse"}}}},"404":{"description":"Asset not found","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApiError"}}}}}}},"/api/v1/agent-sessions/incidents/{incidentId}":{"get":{"tags":["Agent sessions"],"summary":"Generate a mock, provider-neutral incident proposal","operationId":"proposeIncidentActions","parameters":[{"name":"incidentId","in":"path","required":true,"schema":{"type":"string","pattern":"INC-\\d{4}-\\d{4}"}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/IncidentAgentProposal"}}}}}}}},"components":{"schemas":{"WorkOrderResponse":{"type":"object","properties":{"id":{"type":"string"},"incidentId":{"type":"string"},"title":{"type":"string"},"description":{"type":"string"},"assetId":{"type":"string"},"location":{"type":"string"},"priority":{"type":"string","enum":["P1","P2","P3","P4"]},"status":{"type":"string","enum":["Draft","Ready","In Progress","Done"]},"createdAt":{"type":"string","format":"date-time"}}},"WorkOrderWorkflowResponse":{"type":"object","properties":{"outcome":{"type":"string"},"workOrder":{"$ref":"#/components/schemas/WorkOrderResponse"}}},"DraftWorkOrderRequest":{"type":"object","properties":{"incidentId":{"type":"string","example":"INC-2026-0001","minLength":1,"pattern":"INC-\\d{4}-\\d{4}"}},"required":["incidentId"]},"ApiError":{"type":"object","properties":{"timestamp":{"type":"string","format":"date-time"},"status":{"type":"integer","format":"int32"},"error":{"type":"string"},"message":{"type":"string"},"path":{"type":"string"},"fields":{"type":"array","items":{"$ref":"#/components/schemas/ApiFieldError"}}}},"ApiFieldError":{"type":"object","properties":{"field":{"type":"string"},"message":{"type":"string"}}},"CreateIncidentRequest":{"type":"object","properties":{"title":{"type":"string","example":"Transformer smoke detected","maxLength":120,"minLength":0},"description":{"type":"string","example":"Smoke and a burnt smell were reported near the north tunnel transformer room.","maxLength":1000,"minLength":0},"location":{"type":"string","example":"North Tunnel · KM 3.0","maxLength":160,"minLength":0},"assetId":{"type":"string","example":"TRF-NT-003","maxLength":40,"minLength":0},"severity":{"type":"string","enum":["Low","Medium","High","Critical"],"example":"Critical"},"priority":{"type":"string","enum":["P1","P2","P3","P4"],"example":"P1"},"operationalSignals":{"type":"array","example":["Smoke detected","Lighting unavailable","Traffic active"],"items":{"type":"string","maxLength":120,"minLength":0}}},"required":["assetId","description","location","priority","severity","title"]},"IncidentResponse":{"type":"object","properties":{"id":{"type":"string"},"title":{"type":"string"},"description":{"type":"string"},"location":{"type":"string"},"assetId":{"type":"string"},"reportedAt":{"type":"string","format":"date-time"},"severity":{"type":"string","enum":["Low","Medium","High","Critical"]},"priority":{"type":"string","enum":["P1","P2","P3","P4"]},"status":{"type":"string","enum":["Open","Acknowledged","In Progress","Resolved"]},"operationalSignals":{"type":"array","items":{"type":"string"}},"version":{"type":"integer","format":"int64"}}},"TokenResponse":{"type":"object","properties":{"tokenType":{"type":"string"},"accessToken":{"type":"string"},"roles":{"type":"array","items":{"type":"string"}}}},"LoginRequest":{"type":"object","properties":{"username":{"type":"string","example":"operator","minLength":1},"password":{"type":"string","example":"operator","minLength":1}},"required":["password","username"]},"IncidentPageResponse":{"type":"object","description":"One page of the incident search result.","properties":{"content":{"type":"array","description":"Incidents on the current page, ordered by the requested sort.","items":{"$ref":"#/components/schemas/IncidentResponse"}},"page":{"type":"integer","format":"int32","description":"Zero-based index of the current page.","example":0},"size":{"type":"integer","format":"int32","description":"Requested page size.","example":20},"totalElements":{"type":"integer","format":"int64","description":"Total number of incidents matching the filters.","example":42},"totalPages":{"type":"integer","format":"int32","description":"Total number of pages for the requested size.","example":3}}},"AssetResponse":{"type":"object","properties":{"id":{"type":"string"},"name":{"type":"string"},"type":{"type":"string"},"location":{"type":"string"},"criticality":{"type":"string","enum":["Low","Medium","High","Critical"]},"status":{"type":"string","enum":["Operational","Degraded","Maintenance","Out of Service"]},"lastInspectedAt":{"type":"string","format":"date-time"}}},"IncidentAgentProposal":{"type":"object","properties":{"sessionId":{"type":"string"},"provider":{"type":"string"},"mode":{"type":"string"},"incidentId":{"type":"string"},"recommendations":{"type":"array","items":{"$ref":"#/components/schemas/Recommendation"}},"renderBlocks":{"type":"array","items":{"$ref":"#/components/schemas/RenderBlock"}},"tools":{"type":"array","items":{"$ref":"#/components/schemas/ToolContract"}},"guardrails":{"type":"array","items":{"type":"string"}}}},"Recommendation":{"type":"object","properties":{"id":{"type":"string"},"priority":{"type":"string"},"title":{"type":"string"},"summary":{"type":"string"},"requiresApproval":{"type":"boolean"}}},"RenderBlock":{"type":"object","properties":{"type":{"type":"string"},"title":{"type":"string"},"body":{"type":"string"}}},"ToolContract":{"type":"object","properties":{"name":{"type":"string"},"permission":{"type":"string"},"description":{"type":"string"}}}},"securitySchemes":{"bearerAuth":{"type":"http","scheme":"bearer","bearerFormat":"JWT"}}}}